clean up and secure the application

+++ src/resources/v1/status.py
@@ -15,31 +15,43 @@ class Status(Base):
     @marshal_with(DataResponse.marshallable())
     @cache.cached(timeout=60)
     def get(self):
-        auth = request.headers.get('Authorization')
-        if not auth:
-            self.logger.error('Missing authorization')
-            self.throw_error(self.code.BAD_REQUEST)
-
-        auth_token = auth.split(" ")[1]
-
-        is_active = UserTokenModel.is_active(token=auth_token)
-        if not is_active:
-            self.logger.info('Invalid token. Please log in again.')
-            self.throw_error(self.code.UNAUTHORIZED)
-
-        token = UserTokenModel.decode_token(token=auth_token)
         try:
-            uuid.UUID(token['sub'])
-        except ValueError:  # Unknown Error
-            self.logger.info(token)
+            auth = request.headers.get('Authorization')
+            if not auth:
+                raise Exception('Missing authorization')
+            auth_token = auth.split(" ")[1]
+        except Exception as e:
+            self.logger.error(e)
             self.throw_error(self.code.BAD_REQUEST)
-        except TypeError:
-            UserTokenModel.destroy_auth_token(auth_token=auth_token)  # destroy the token
-            self.logger.info(token)
+
+        try:
+            is_active = UserTokenModel.is_active(token=auth_token)
+            if not is_active:
+                raise Exception('Invalid token. Please log in again.')
+        except Exception as e:
+            self.logger.error(e)
             self.throw_error(self.code.UNAUTHORIZED)
 
-        user = UserModel.query.filter(UserModel.uuid == token['sub']).first()
+        try:
+            token = UserTokenModel.decode_token(token=auth_token)
+        except ValueError as e:
+            if 'destroy_token' in e.args:
+                UserTokenModel.destroy_auth_token(auth_token=auth_token)  # destroy the token
+            self.logger.error(e)
+            self.throw_error(self.code.UNAUTHORIZED)
 
-        user_result = UserSchema().dump(user)
+        # query auth_db for user in request payload
+        try:
+            user = UserModel.query.filter(UserModel.uuid == token['sub']).first()
+        except Exception as e:
+            self.logger.error(e)
+            self.throw_error(self.code.INTERNAL_SERVER_ERROR)
+
+        # dump user model instance
+        try:
+            user_result = UserSchema().dump(user)
+        except Exception as e:
+            self.logger.error(e)
+            self.throw_error(self.code.INTERNAL_SERVER_ERROR)
 
         return DataResponse(data={'user': user_result})
